Move cast code to a module
