+++ Makefile
@@ -8,6 +8,7 @@ INCDIRS := include
 # Project files
 SRCS := $(addprefix $(SRCDIR)/,\
 	main.c \
+	cast/cast.c \
 	input/input.c \
 	input/input_funcs.c \
 )
